display shipping informations in orderDetailsPage

+++ src/components/Admin/ListScreen/ListItem/ProdListItem.js
@@ -6,33 +6,37 @@ import ListItem from './ListItem';
 import ProdDescription from './../../../ListItems/ProdDescription/ProdDescription';
 
 const ProdListItem = (props) => {
-	const tools = [
-		{
-			icon: <FiEye />,
-			handler: () => props.additional.viewHandler(props.itemData),
-		},
-		{
-			icon: <FiEdit />,
-			handler: () =>
-				props.additional.modifyHandler(props.itemData, 'modify'),
-		},
-		{
-			type: 'delete',
-			icon: <FiTrash2 />,
-			handler: () =>
-				props.additional.modifyHandler(props.itemData, 'remove'),
-		},
-	];
+	let tools = null;
+	let isRemoved;
+	let isModified;
 
-	const isRemoved = isPresent(
-		props.itemData.id,
-		props.additional.removedItems
-	);
+	if (props.additional) {
+		tools = [
+			{
+				icon: <FiEye />,
+				handler: () => props.additional.viewHandler(props.itemData),
+			},
+			{
+				icon: <FiEdit />,
+				handler: () =>
+					props.additional.modifyHandler(props.itemData, 'modify'),
+			},
+			{
+				type: 'delete',
+				icon: <FiTrash2 />,
+				handler: () =>
+					props.additional.modifyHandler(props.itemData, 'remove'),
+			},
+		];
 
-	const isModified = isPresent(
-		props.itemData.id,
-		props.additional.modifiedItems
-	);
+		isRemoved = isPresent(props.itemData.id, props.additional.removedItems);
+
+		isModified = isPresent(
+			props.itemData.id,
+			props.additional.modifiedItems
+		);
+	}
+
 	return (
 		<ListItem removed={isRemoved} edited={isModified} buttons={tools}>
 			<ProdDescription itemData={props.itemData} />
